+++ src/main/java/App/Taller.java
@@ -9,16 +9,16 @@ import org.jgrapht.graph.DefaultEdge;
 public class Taller {
     public String nombre;
     public String sistema_asociado;
-    public int dinero_fallas_menores;
+    public long dinero_fallas_menores;
     public static Hashtable<String, DefaultEdge> TallerNombres = new Hashtable<>();
     public static TreeMap<String, ArrayList<String>> TallerSistemas = new TreeMap<>();
-    public static TreeMap<Integer, ArrayList<String>> TallerDinero = new TreeMap<>();
+    public static TreeMap<Long, ArrayList<String>> TallerDinero = new TreeMap<>();
 
 
     public Taller(String nombre, String sistema_asociado, String dinero_fallas_menores) {
         this.nombre = nombre.toLowerCase();
         this.sistema_asociado = sistema_asociado.toLowerCase();
-        this.dinero_fallas_menores = Integer.parseInt(dinero_fallas_menores);
+        this.dinero_fallas_menores = Long.parseLong(dinero_fallas_menores);
         App.Grafo.addVertex(this);
         TallerNombres.put(this.nombre,App.Grafo.addEdge(this, this));
         
@@ -29,12 +29,12 @@ public class Taller {
             lista.add(this.nombre);
             TallerSistemas.put(sistema_asociado.toLowerCase(), lista);
         }
-        if (TallerDinero.containsKey(Integer.parseInt(dinero_fallas_menores))) {
-            TallerDinero.get(Integer.parseInt(dinero_fallas_menores)).add(this.nombre);
+        if (TallerDinero.containsKey(Long.parseLong(dinero_fallas_menores))) {
+            TallerDinero.get(Long.parseLong(dinero_fallas_menores)).add(this.nombre);
         } else {
             ArrayList<String> lista = new ArrayList<>();
             lista.add(this.nombre);
-            TallerDinero.put(Integer.parseInt(dinero_fallas_menores), lista);
+            TallerDinero.put(Long.parseLong(dinero_fallas_menores), lista);
         }
 
     }
